+++ js/app.js
@@ -1,33 +1,6 @@
 var React = require('react');
-
-
-var ObjectRow = React.createClass({
-    render: function() {
-        return (
-            <img src="static/bananas.png"></img>
-        )
-    }
-});
-
-var BananaComponent = React.createClass({
-    render: function() {
-        var rows = [];
-        for (var i=0; i < this.props.number; i++) {
-            rows.push(<ObjectRow />);
-        }
-        return (<tbody>{rows}</tbody>);
-    }
-});
-
-var AppleComponent = React.createClass({
-    render: function() {
-        return (
-            <img src="static/apple.png"></img>
-        )
-    }
-});
-
-
+var BananaComponent = require('./components/Banana.react');
+var AppleComponent = require('./components/Apple.react');
 
 var FruitComponent = React.createClass({
     render: function() {
