+++ js/components/Apple.react.js
@@ -1,0 +1,9 @@
+var AppleComponent = React.createClass({
+    render: function() {
+        return (
+            <img src="static/apple.png"></img>
+        )
+    }
+});
+
+module.exports = AppleComponent;
