+++ js/components/Banana.react.js
@@ -1,0 +1,19 @@
+var ObjectRow = React.createClass({
+    render: function() {
+        return (
+            <img src="static/banana.png"></img>
+        )
+    }
+});
+
+var BananaComponent = React.createClass({
+    render: function() {
+        var rows = [];
+        for (var i=0; i < this.props.number; i++) {
+            rows.push(<ObjectRow />);
+        }
+        return (<tbody>{rows}</tbody>);
+    }
+});
+
+module.exports = BananaComponent;
