+++ package.json
@@ -7,7 +7,8 @@
   "dependencies": {
     "classnames": "^2.1.3",
     "flux": "^2.0.0",
-    "react": "^0.13.0"
+    "react": "^0.13.0",
+    "webpack": "^1.8.5"
   },
   "devDependencies": {
     "browserify": "^6.2.0",
